Move the GestureRecognizer feature configuration flags to the right layer
